#1 Removed unused stuff. Moved mi types.
