+++ code/teltonika-demo/teltonika-demo-web-server/src/main/java/fielden/webapp/WebUiConfig.java
@@ -7,14 +7,14 @@ import org.apache.commons.lang3.StringUtils;
 import fielden.config.Modules;
 import fielden.config.personnel.PersonWebUiConfig;
 import fielden.main.menu.personnel.MiTgJourney;
+import fielden.main.menu.personnel.MiTgMachineRealtimeMonitor;
+import fielden.main.menu.personnel.MiTgMessage;
+import fielden.main.menu.personnel.MiTgPolygon;
+import fielden.main.menu.personnel.MiTgStop;
 import fielden.personnel.Person;
 import ua.com.fielden.platform.basic.config.Workflows;
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.reflection.TitlesDescsGetter;
-import ua.com.fielden.platform.ui.menu.sample.MiTgMachineRealtimeMonitor;
-import ua.com.fielden.platform.ui.menu.sample.MiTgMessage;
-import ua.com.fielden.platform.ui.menu.sample.MiTgPolygon;
-import ua.com.fielden.platform.ui.menu.sample.MiTgStop;
 import ua.com.fielden.platform.utils.Pair;
 import ua.com.fielden.platform.web.app.config.IWebUiBuilder;
 import ua.com.fielden.platform.web.interfaces.ILayout.Device;
