+++ code/teltonika-demo/teltonika-demo-web-ui/src/main/java/fielden/main/menu/personnel/MiTgStop.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.ui.menu.sample;
+package fielden.main.menu.personnel;
 
 import ua.com.fielden.platform.entity.annotation.EntityType;
 import ua.com.fielden.platform.sample.domain.TgStop;
